Add ability to inject properties in non public classes
Until now, Easy Props relied on Apache commons-beanutils
to set properties on target objects. This requires the
object class to be public and to provide a public setter.
I believe this is because Apache commons-beanutils is
targeted at JavaBeans (where a default constructor +
public getters/setters are required).

This commit use the Java reflection API to set the value
directly on the field. This is less restrictive and makes
it possible to inject properties in private classes without
requiring a public setter for the target field.

Issue #8